Move thread_state to style.
