+++ components/style/workqueue.rs
@@ -18,8 +18,8 @@ use deque::{self, Abort, Data, Empty, Stealer, Worker};
 use rand::{Rng, XorShiftRng, weak_rng};
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::mpsc::{Receiver, Sender, channel};
+use thread_state;
 use util::thread::spawn_named;
-use util::thread_state;
 
 /// A unit of work.
 ///
